Clean up Playwright config and spec files
